use wrapped enum everywhere
