+++ python/basix/finite_element.py
@@ -5,12 +5,13 @@ import typing
 import numpy as _np
 import numpy.typing as npt
 
-from basix._basixcpp import CellType, DPCVariant, ElementFamily
+from basix._basixcpp import DPCVariant, ElementFamily
 from basix._basixcpp import FiniteElement_float32 as _FiniteElement_float32  # type: ignore
 from basix._basixcpp import FiniteElement_float64 as _FiniteElement_float64  # type: ignore
 from basix._basixcpp import LagrangeVariant, MapType, PolysetType, SobolevSpace
 from basix._basixcpp import create_custom_element as _create_custom_element
 from basix._basixcpp import create_element as _create_element  # type: ignore
+from basix.cell import CellType
 
 __all__ = ["FiniteElement", "create_element", "create_custom_element", "string_to_family"]
